move indexing functions to top-level conftest
